refactor(types): improve namespace migration.
